Centralize the log adapter
This should eventually be updated to map the levels to the framework
logger and appropriate module-print_* function.
