+++ lib/msf/core/exploit/remote/smb/server.rb
@@ -5,6 +5,7 @@ module Msf
     # This mixin provides a minimal SMB server
     module Server
       include ::Msf::Exploit::Remote::SocketServer
+      include ::Msf::Exploit::Remote::SMB::LogAdapter
 
       module ServiceMixin
         def start
@@ -54,9 +55,15 @@ module Msf
             }
         )
 
+        unless opts[:logger]
+          log_device = LogAdapter::LogDevice::Framework.new(framework)
+          opts[:logger] = LogAdapter::Logger.new(self, log_device)
+        end
+
         server = RubySMB::Server.new(
           server_sock: @rsock,
-          gss_provider: opts[:gss_provider]
+          gss_provider: opts[:gss_provider],
+          logger: opts[:logger]
         )
 
         server.extend(ServiceMixin)
